requirements.txt
fastapi==0.103.1
google-cloud-logging==3.9.0
google-cloud-storage==2.14.0
google-cloud-bigquery
retry==0.9.2
requests-toolbelt==1.0.0
psutil